// src/utils/toast/toast.d.ts
export interface Options {
  class?: string | string[]
  disableClick?: boolean
  duration?: number | false
  pauseOnHover?: boolean
  max?: number
  positionX?: 'left' | 'right' | 'center'
  positionY?: 'top' | 'bottom'
  styles?: Record<string, string>
}
// ...
export interface LocalOptions {
  class?: string | string[]
  disableClick?: boolean
  duration?: number | false
  pauseOnHover?: boolean
  link?: { href: string; targetBlank?: boolean }
  positionX?: 'left' | 'right' | 'center'
  positionY?: 'top' | 'bottom'
  slot?: string
  slotLeft?: string
  slotRight?: string
  styles?: Record<string, string>
  type?: 'error' | 'success' | 'passive'
}
// ...
export type TRenderFunction = (
  text: string,
  options: LocalOptions | undefined
) => void